requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
transformers==4.43.3
sentence-transformers==3.0.1
torch==2.3.1
scikit-learn==1.5.1
numpy==1.26.4
nltk==3.8.1
pymupdf==1.24.9
python-docx==1.1.2
pydantic==2.8.2
python-multipart==0.0.9
streamlit==1.37.0
flask==3.0.3
jinja2==3.1.4
tqdm==4.66.4 
ipykernel
